+++ src/plone/app/multilingual/tests/test_lrf.py
@@ -7,7 +7,7 @@ from plone.app.multilingual.testing import PAM_FUNCTIONAL_TESTING
 from plone.dexterity.utils import createContentInContainer
 from plone.uuid.interfaces import IUUID
 from zope.event import notify
-import unittest2 as unittest
+import unittest
 from plone.app.multilingual.interfaces import IPloneAppMultilingualInstalled
 from zope.interface import alsoProvides